# extras/4_Head_poltary_system/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/purchases": {
      "post": {
        "operationId": "PurchasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new purchase (DRAFT)",
        "tags": [
          "Purchases"
        ]
      },
      "get": {
        "operationId": "PurchasesController_findAll",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all purchases",
        "tags": [
          "Purchases"
        ]
      }
    },
...
    "/purchases/{id}": {
      "get": {
        "operationId": "PurchasesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get purchase by ID",
        "tags": [
          "Purchases"
        ]
      },
      "patch": {
        "operationId": "PurchasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update purchase (only DRAFT)",
        "tags": [
          "Purchases"
        ]
      },
      "delete": {
        "operationId": "PurchasesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete purchase (only DRAFT)",
        "tags": [
          "Purchases"
        ]
      }
    },
...
    "/sales": {
      "post": {
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new sale (DRAFT)",
        "tags": [
          "Sales"
        ]
      },
      "get": {
        "operationId": "SalesController_findAll",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all sales",
        "tags": [
          "Sales"
        ]
      }
    },
...
    "/sales/{id}": {
      "get": {
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get sale by ID",
        "tags": [
          "Sales"
        ]
      },
      "patch": {
        "operationId": "SalesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update sale (only DRAFT)",
        "tags": [
          "Sales"
        ]
      },
      "delete": {
        "operationId": "SalesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete sale (only DRAFT)",
        "tags": [
          "Sales"
        ]
      }
    },
...
    "/reports/profit-loss": {
      "get": {
        "operationId": "ReportingController_getProfitLoss",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reporting"
        ]
      }
    },
...
    "/departments/supply/profit-loss": {
      "get": {
        "operationId": "SupplyController_getProfitLoss",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Supply"
        ]
      }
    },
...
    "/departments/wastage/profit-loss": {
      "get": {
        "operationId": "WastageController_getProfitLoss",
        "parameters": [
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wastage"
        ]
      }
    },
...
  },
  "info": {
    "title": "LabVerse API",
    "description": "Complete project management and CRM system API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePurchaseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePurchaseDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}